+++ CHANGELOG.md
@@ -7,6 +7,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- `AsyncClient`: Prevent deadlocks when dropping instances within an asynchronous context. This only
+  works when the feature `"tokio"` is enabled.
+
 ## [0.10.0] - 2025-10-14
 
 ### Added